+++ src/balance/main.ts
@@ -1,26 +1,24 @@
-import Entropy from "@entropyxyz/sdk"
-import { EntropyBase } from "../common/entropy-base"
 import * as BalanceUtils from "./utils"
 import { BalanceInfo } from "./types"
+import { EntropySubstrateBase } from "src/common/entropy-substrate-base"
 
 const FLOW_CONTEXT = 'ENTROPY-BALANCE'
-export class EntropyBalance extends EntropyBase {
-  constructor (entropy: Entropy, endpoint: string) {
-    super({ entropy, endpoint, flowContext: FLOW_CONTEXT })
+export class EntropyBalance extends EntropySubstrateBase {
+  constructor (substrate: any, endpoint: string) {
+    super({ substrate, endpoint, flowContext: FLOW_CONTEXT })
   }
 
-  async getBalance (address: string): Promise<number> {
-    const accountInfo = (await this.entropy.substrate.query.system.account(address)) as any
+  async getAnyBalance (address: string) {
+    const accountInfo = (await this.substrate.query.system.account(address)) as any
     const balance = parseInt(BalanceUtils.hexToBigInt(accountInfo.data.free).toString())
 
-    this.logger.log(`Current balance of ${address}: ${balance}`, EntropyBalance.name)
     return balance
   }
 
   async getBalances (addresses: string[]): Promise<BalanceInfo[]> {
     return Promise.all(
       addresses.map(async address => {
-        return this.getBalance(address)
+        return this.getAnyBalance(address)
           .then((balance: number) => {
             return { address, balance }
           })