+++ src/transfer/main.ts
@@ -1,14 +1,16 @@
-import Entropy from "@entropyxyz/sdk";
+// @ts-ignore
+import { Pair } from '@entropyxyz/sdk/keys'
 
-import { EntropyBase } from "../common/entropy-base";
 import { bitsToLilBits, formatDispatchError, getTokenDetails } from "../common/utils";
+
 import { TransferOptions } from "./types";
+import { EntropySubstrateBase } from 'src/common/entropy-substrate-base';
 
 const FLOW_CONTEXT = 'ENTROPY_TRANSFER'
 
-export class EntropyTransfer extends EntropyBase {
-  constructor (entropy: Entropy, endpoint: string) {
-    super({ entropy, endpoint, flowContext: FLOW_CONTEXT })
+export class EntropyTransfer extends EntropySubstrateBase {
+  constructor (substrate: any, endpoint: string) {
+    super({ substrate, endpoint, flowContext: FLOW_CONTEXT })
   }
 
   // NOTE: a more accessible function which handles
@@ -16,15 +18,17 @@ export class EntropyTransfer extends EntropyBase {
   // - converting `amount` (string => BigInt)
   // - progress callbacks (optional)
 
-  async transfer (toAddress: string, amountInBits: string) {
-    const { decimals } = await getTokenDetails(this.entropy)
+  async transfer (from: Pair, toAddress: string, amountInBits: string) {
+    const { decimals } = await getTokenDetails(this.substrate)
     const lilBits = bitsToLilBits(Number(amountInBits), decimals)
 
-    return this.rawTransfer({
-      from: this.entropy.keyring.accounts.registration.pair,
+    const transferStatus = await this.rawTransfer({
+      from,
       to: toAddress,
       lilBits
     })
+
+    return transferStatus
   }
 
   private async rawTransfer (payload: TransferOptions): Promise<any> {
@@ -33,12 +37,12 @@ export class EntropyTransfer extends EntropyBase {
     return new Promise((resolve, reject) => {
       // WARN: await signAndSend is dangerous as it does not resolve
       // after transaction is complete :melt:
-      this.entropy.substrate.tx.balances
+      this.substrate.tx.balances
         .transferAllowDeath(to, lilBits)
         // @ts-ignore
         .signAndSend(from, ({ status, dispatchError }) => {
           if (dispatchError) {
-            const error = formatDispatchError(this.entropy, dispatchError)
+            const error = formatDispatchError(this.substrate, dispatchError)
             this.logger.error('There was an issue sending this transfer', error)
             return reject(error)
           }