+++ tests/testing-utils/constants.mjs
@@ -10,4 +10,5 @@ export const eveSeed =
   '0x786ad0e2df456fe43dd1f91ebca22e235bc162e0bb8d53c633e8c85b2af68b7a'
 export const eveAddress = '5HGjWAeFDfFCWPsjFQdVV2Msvz2XtMktvgocEZcCj68kUMaw'
 
+export const DEFAULT_ENDPOINT = 'ws://127.0.0.1:9944'
 export const DEFAULT_TOKEN_DECIMALS = 10